aio-smoosh-struct-kiocb-fix
fs/aio.c: In function 'io_submit_one':
fs/aio.c:1256: warning: cast to pointer from integer of different size

Cc: Al Viro <viro@zeniv.linux.org.uk>
Cc: Asai Thambi S P <asamymuthupa@micron.com>
Cc: Benjamin LaHaise <bcrl@kvack.org>
Cc: Felipe Balbi <balbi@ti.com>
Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Cc: Jeff Moyer <jmoyer@redhat.com>
Cc: Jens Axboe <axboe@kernel.dk>
Cc: Joel Becker <jlbec@evilplan.org>
Cc: Kent Overstreet <koverstreet@google.com>
Cc: Mark Fasheh <mfasheh@suse.com>
Cc: Rusty Russell <rusty@rustcorp.com.au>
Cc: Sam Bradshaw <sbradshaw@micron.com>
Cc: Selvan Mani <smani@micron.com>
Cc: Zach Brown <zab@redhat.com>

+++ fs/aio.c
@@ -1253,7 +1253,7 @@ static int io_submit_one(struct kioctx *ctx, struct iocb __user *user_iocb,
 	req->ki_pos		= iocb->aio_offset;
 	req->ki_nbytes		= iocb->aio_nbytes;
 	req->ki_left		= iocb->aio_nbytes;
-	req->ki_buf		= (char __user *) iocb->aio_buf;
+	req->ki_buf		= (char __user *)(unsigned long)iocb->aio_buf;
 	req->ki_nr_segs		= 0;
 	req->ki_cur_seg		= 0;
